requirements.txt
ping3==4.0.4
pyserial==3.5
